fix: 마우스 관련 로직 CanvasObjectComposite class 에서 Model class 로 분리
- 그에 따른 state 일부 변경
